Set OpenStack primarySubnet in MachineSpec
In the OpenStack platform, the primary subnet is where a node's kubernetes endpoints are published.
Right now, we set this with a tag in OpenStack. Due to some users having restrictions on tagging
resources, and this being a difficult user interface, we added a way to modify this value in the
machine api. We still fully support setting this value with tagging for backwards compatibility
reasons. In this patch, we update the installer code to set the primarySubnet using the MachineSpec.