pkg: validation for previously existing ovirt configuration
This change introduces validation of a previously-existing ovirt configuration in ~/.ovirt/ovirt-config.yaml with invalid credentials. Before this change an existing ovirt-config.yaml would cause the installer to fail without asking for credentials. This change adds error handling and will post questions to the user if the credentials are invalid.

Fixes OCPRHV-423