doc/openstack: update quota reqs for kuryr
Update the Neutron quotas now required when Kuryr is used at the
networking plugin for an OCP cluster.

The numbers were taken from:
https://docs.openshift.com/container-platform/4.6/installing/installing_openstack/installing-openstack-user-kuryr.html#installation-osp-default-kuryr-deployment_installing-openstack-user-kuryr

and confirmed with the Kuryr team.

It updates the table and the command that updates the quotas so the
reader won't be confused nor set the wrong quotas, which lead to a
failed deployment of that type: http://paste.openstack.org/show/799839/