Bug 1888464: add tag:UnTagResource perm for aws shared networks
Create a new permissions group for destroying a shared-network cluster.
These clusters require permissions to untag the shared networks.
This check is skipped in the case of C2S, because delete is not supported.

Skip checking delete permissions in case of C2S.

Cluster destroy is not supported for the C2S use case, so we should
not require permissions for delete.