tls: add asset to load user-provided bound-sa keys
Allow users to provide the service account signing keys at
- $INSTALL_DIR/tls/bound-service-account-signing.key
The installer reads the private RSA key and extracts the public key creates the corresponding bound-service-account-signing.pub
file for the bootstrap host.

These files are traferred to the bootstrap host so that kap-o can read these files
and configure the bootstrap-kube-apiserver and in-cluster kube-apiserver with this key.

kap-o render PR https://github.com/openshift/cluster-kube-apiserver-operator/pull/1006 adds the functionality
to read these files from /opt/openshift/tls/<> and setup the bootkube-kube-apiserver to use the key on bootstrap host
and also create a secret with these files so in-cluster kap-o can configure the actual kube-apiserver to also use these
keys.