baremetal: add units to podman wait in ironic
A [bug in some versions of podman](https://bugzilla.redhat.com/show_bug.cgi?id=1897282)
means that units end up being required for the podman wait command, even
though it was documented as optional. This is fixed, but the podman
requiring units shipped in RHEL 8.3 and recent Fedoras, and won't be
updated until 8.4.  This proactively adds units to baremetal's
podman-wait commands in case this broken podman ends up in RHCOS or
FCOS. It doesn't hurt to be explicit anyway.