Bug 1895874: RHV installer now uses the downloaded certificate from the engine
During the installation the installer downloads, displays the oVirt engine certificate to the user and asks for confirmation. Previously, this certificate was not used during the installation and the user had to add the certificate to the trust root. This change stores the certificate in ~/.ovirt/ovirt-config.yaml and uses it during the installation.