Bug 1889779: vSphere destroy: handle failed clusters
Adjusts vSphere destroy code to handle cases when cluster create was
terminated early, particularly if interrupted before OVA is completely
downloaded. This commit moves the tagging of the OVA template before
the data upload, because the template was not getting tagged if the
upload did not commit. See details below.

Also, prior to this change, destroy would stop if no VMs or folders
were found, but this updates the flow to allow best effort to destroy
all resources.

Finally, updates caps to be consistent with other standard logger output.

Import OVA & Destroy Details:
If one hard terminates the installer during a template import, the import
continues in the vCenter for some period of time afterward (I would estimate
at five minutes).

This will block the deletion of the template, but in my testing the code
gracefully handles this and spins until the lock is resolved and then
deletion completes successfully.