libvirt: allow customization of the pool path
When creating the storage pool path, that path needs to be on a file
system big enough to accommodate virtual machine images spun up by the
installer. In some cases the default `/var/lib/libvirt/openshift-images`
could reside on a file system that if filled might cause issues, for
example when `/var` resides on the root file system, as it does by
default on Fedora Linux.

This allows customizing the libvirt storage pool directory path,
allowing the user to choose where to store the virtual machine images.

The default is left to `/var/lib/libvirt/openshift-images`.