baremetal: Add ironic logwatch containers
A new entrypoint has been added to the ironic containers
that watches for deploy ramdisk logs in the shared volume
and exposes them via the regular podman logs output.

https://github.com/openshift/ironic-inspector-image/pull/47/
https://github.com/openshift/ironic-image/pull/121/

This should make it easier for users debugging as they no longer
need to exec into the containers and have knowledge of the
internal logfile location.