Fixes OCVRHV-45: Allow preallocated disk
This change adds the option to use full clones instead of thin provisioning for the master nodes. Full provisioning increases the performance at the cost of up-front disk usage. The default is now on, but can be turned off by the user in the survey.

The install config can be set to the old behavior (thin provisioning) by setting the `ovirt_master_clone` option to `false`.