BZ 1904125: Ensure the bootstrap ignition libvirt pool defaults to `<clustername>.<id>.bootstrap` rather than `default`
Currently in the 'baremetal' platform, we use a customized TF-created storage volume with the following naming convention: `<clustername>.<id>.bootstrap` and this volume is used to store the bootstrap VM images. Since the libvirt_ignition resources was introduced, it has been defaulting to the `default` libvirt storage pool.

This PR ensures that the libvirt_ignition TF resource uses the above mentioned customized libvirt storage pool.