Bug 1906459: openstack: fix quota checks when they're unlimited
In OpenStack, if a quota resource is limited to "-1", it means that it's
unlimited. It's the case of a lot of OpenStack clouds, were quotas
aren't defined for all projects and by default "-1" is used.

This patch will ignore a quota record if its limit is set to "-1".

Note: it adds "Unlimited" to the Quota struct, as a boolean, so later
any platform can set it to true if a resource is unlimited.

Signed-off-by: Emilien Macchi <emilien@redhat.com>