Bug 1903055: OpenStack: move platform validations to cluster asset
Now we perform validations of install-config asset right after
its generation. Unfortunately, later this asset is modified
during "machines" generation. It leads to the fact that even
if the validation succeeds, installation may fail after that.

To prevent this we move the validation to "cluster" asset,
because there are no install-config modifications after that.

Also this commit removes defaultMachinePlatform validation, because
it's not used for the cluster provisioning and just contains default
values for the real machine pools.