installconfig: enable vsphere folders for datacenters and clusters.
This change modifies the vsphere prompt behavior for both datacenter and
cluster inputs. When these items exist within one or more folders in
the vsphere infrastructure, the folder names will be added as a prefix
to the item in the list of choices. As a result, this folder information
will now be saved into the install-config.yaml and used to deploy the
cluster.