Bug 1908171: fix Terraform issue with GCP custom machine types
Similar to previous bug (see below) but in this case for custom
machine types. In the previous bz, the config generated by Terraform
did not match the resource that was ultimately created. So when
apply is run a second time for bootstrap destroy, an error occurs
when TF tries to switch the resource back to match the config.

Similar: fdefcca25cb12fc241198b0688f421411c9a580a