Make dual stack on Azure require an environment variable override
In OCP 4.8 we don't officially support dual stack on Azure, but it
appears to _mostly_ work, and we will be using it for some CI and
development purposes. Add an environment variable
(OPENSHIFT_INSTALL_EXPERIMENTAL_DUAL_STACK=true) so developers/CI can
use it, but customers don't accidentally use it thinking it's
supported.