Set same additional networks on Bootstrap as Control Plane
The bootstrap node was only being attached to the private network created by the installer,
causing deployments with more advanced networking topologies that made use of the additional
networking feature to fail. The bootstrap needs to be attached to the same networks as the
control plane nodes in order to guarantee the install will succeed.

Fixes Bug 1908583