Don't store Flavor as an encapsulated pointer
This prevents a potential usage error where a returned Flavor is
non-nil, but usage of it causes a nil pointer dereference because the
encapsulated Flavor is nil. It is a non-functional change, because the
collectInfo should not be able to set a nil pointer here anyway.

This change simplifies the check for flavor existence: the flavor is
valid if it has a key in ci.Flavors.