Set Additional Control Plane Security Groups on Bootstrap Node
The bootstrap node in the installer does not have any way to set additional
networks and security groups. The installer currently adds additional control
plane networks to the bootrap node, and should do the same for security groups.
Since the bootstrap is an ephemeral node that has the same networking requirements
as the master nodes, this logically makes sense and makes the user experience more
consistent.

Fixes: Bug 1915998