Bug 1916692: OpenStack: Delete leftover LBs when destroying cluster
If the user destroyed a cluster without removing all the associated
service LBs, the `destroy` command would fail to remove the network
and loop until it hits the timeout.

The destroy command now looks if there are any leftover LBs where its
`VipNetworkID` matches the network ID and deprovisions it.  We filter on
services LBs created by the openstack cloud provider, matching the
`Kubernetes external service` string in the description [1], to ensure
we're not destroying a user-created resource by mistake.

[1] https://github.com/openshift/kubernetes/blob/442a69c/staging/src/k8s.io/legacy-cloud-providers/openstack/openstack_loadbalancer.go#L446