openstack/validation: enforce control plane size
This is a cherry-pick of commit fb94d1cd4d713ee79efa672af71b52a067cb2e36

This is a follow up to b6e3088d, which hard codes the size of the
control plane to three nodes. We have a broad policy within OpenShift to
only support three-node control planes, but that's there's nothing
technically preventing a user from using a different size. In the case
of IPI on OpenStack, however, there is a technical restriction, so this
explicitly validates that.