data/bootstrap: Check if release image architecture matches host architecture
This is a simple sanity check to see if release image architecture matches the bootstrap node's architecture. This is useful in cases of
user error where an incorrect release image is specified and we error right away rather than downloading the associated images and failing in
say crio when downloading the incorrect pause image.