Fix cluster destroy when byo is used with Kuryr
When using byo the cluster destroy was relying on the
primary network tag to identify the router used and remove
the extra interfaces. As the provided network is not tagged
anymore the cluster is not able to get cleaned up. This commit
fixes the issue by relying on the network used by the Servers
to identify the router used.