Bug 1926558: rhcos.json: Update bootimages for ppc64le/s390x to fix kubelet erroring out with "No authentication method configured"
Refer to https://bugzilla.redhat.com/show_bug.cgi?id=1926558. After https://github.com/openshift/installer/pull/4604 landed in 4.4, the multi-arch CI runs started failing because the hyperkube package was old. This was not an issue for x86 as the rhcos images were bumped for a CVE (https://github.com/openshift/installer/pull/3985), resulting in a newer hyperkube package which did not have this issue with the kubelet erroring out on `anonymous-auth=false`

These are the builds that were closest in parity to x86:
https://releases-art-rhcos.svc.ci.openshift.org/art/storage/releases/rhcos-4.4-s390x/44.82.202011122341-0/s390x/meta.json for s390x
https://releases-art-rhcos.svc.ci.openshift.org/art/storage/releases/rhcos-4.4-ppc64le/44.82.202012010439-0/ppc64le/meta.json for ppc64le