proxy: Validate noProxy input and add prefix for proxy urls
The installer currently trims the spaces for the noProxy URLs
provided by the user but the Cluster Network Operator does not which
causes the machine config operator to fail with invalid proxy urls.
Adding a validation check to reject the input from the user if there
are spaces in the noProxy input.

Also added validation checks for the HTTPProxy and the HTTPSProxy similar to
the cluster network operator validation done for the same.