machines/openstack: re-enable external clientOpts
In 05453ef0d the MachineSets function was updated to remove the option
to specify external clientOpts. This clientOpts is used by hive to generate
the machinesets and ensuring the correct client authentication is used.

using the default client opts as forced in 05453ef0d only allows configuring the
openstack client using the default location on the host, which doesn't sit well
with something like hive which needs to manage the machinesets for many different
clients from single location.

Reverting the change in 05453ef0d to allow machinesets to accept external clientOpts
and use the default opts when it is nil.