Add bootstrapInPlace to insall config
Added new asset for SingleNodeBootstrapInPlace

Updated bootkube to set BOOTSTRAP_INPLACE variable and evaluate it at run time.
This change allow someone looking at the script to follow the shell logic for
single-node vs. multi-node deployment while debugging.

When creating single-node-ignition-config we now validate
that the install-config contain configuration for bootstrapInPlace.
Added install-to-disk serivce that will complete the installation
by writing the OS to the desired instnallation disk and reboot the node.

Signed-off-by: Eran Cohen <eranco@redhat.com>