images: Fix GCP openssl not found error
With the current code, yum install pyopenssl is used to install
the openssl package that gcloud sdk needs to signurl which has
stopped working because gcloud is not able to pick up the package.

Added a code to install pip of a version less than 21 (pip support
for python 2 ends with pip version 21) and installed pyopenssl
which is the recommended way to install openssl for gcloud.