gather-bootstrap: Gather logs before bootstrap shuts down
The bootkube tries to keep the bootstrap control-plane on for
30 minutes before it shuts down due to a timeout. In which case,
the bootkube tries to re-run the control-plane to gather the logs.
The re-run is not guaranteed by the api team and hence there is
a possibility the logs for the control-plane will not be gathered.

Adding a check to trap the error message from the run bootstrap
command in the bootkube to run installer-gather script to gather
the logs in case of any error before the resources get deleted.
The logs will be stored in a folder with a default name temporarily.
Then the size of the openapi.json.gz file will be checked to see
if there was any extra information was logged from the control-
plane and if there was not a change, the tar file and the folder
will be removed as the bootstrap control-plane did not come up.
If it has been changed, the folder is moved from temporary to
permanent folder (removing the temp part of the name).

Also adding a mechanism to pull the created logs with the default
folder name to get the latest information about the control-plane.