baremetal: always use API VIP for installer communication
The installer's terraform provider needs to talk to Ironic running on
the bootstrap VM. That communication path always used the provisioning
network, which means when using Hive from a remote cluster, the
provisioning network must be routable.

Since 4.7, with the disabled provisioning network feature, the installer
uses the API VIP for installer communication and image cache. It makes
sense to remove the dependency that the installer talks to the
provisioning network at all.  This PR makes the installer always use the
API VIP to talk to Ironic, but maintains the provisioning network for
bare metal host <-> image cache communications when it's enabled.