Bug 1903055: Set default values to machine pools before validation
Now we validate just the user input, but in fact we additionally
set default values there if they are not provided directly. It leads
to incorrect validation of the install config.

This commit sets the default values to machine pools before validation
to make them just like what we use to build a cluster.