gcp: Ignore Licenses: vmx, only use public image
The history here is basically that originally for GCP, we
cloned a disk image, just like how Azure works today.

Then later, GCP cracked down on the allowable frequency of image
creation and told us to use public bootable images, so we started
uploading a public RHCOS base image.

But we originally did that without the nested virt license, so
the KubeVirt team added a PR to clone the image and add the
nested virt license.

And then finally, in 4.6 we fixed the CoreOS public image
to add the nested virt license by default (which also, GCP
recommended we do):
https://github.com/coreos/coreos-assembler/pull/1618/commits/73320cf633e362f96de5f9d9aabb02b8647d4156

So there's no need for any of this image cloning or license
stuff anymore.  For compatibility (with e.g. the KubeVirt CI)
we silently ignore the `enable-vmx` licenses because it's
implicitly enabled by default anyways.  Copying the image
to enable it is a pessimization in fact!

Other licenses are a fatal error - I am not aware of
any that anyone would want.  And anyone who wants to
do something truly custom can use
the image override variable anyways; they don't *have*
to use our Terraform code.  Or they can hack the installer
code.

Prep for using the stream metadata in
https://github.com/openshift/installer/pull/4582