Baremetal: Explicitly set watchAllNamespaces
A new value has been added to the baremetal provisioning CRD that allows
it to watch more namespaces to support deploying remote single node
OpenShift instances.  This patch explicitly sets this value to false.

We do not want to expose this in the installer at this time.  This is
a day 2 option.