Revert "baremetal: send full ignition to masters"
This doesn't work for IPI baremetal deployments driven via hive,
because there are firewall rules that prevent access to the
bootstrap MCS from the pod running the installer.

This was implemented in:
https://github.com/openshift/installer/pull/4427

But we ran into problems making the same approach work for
worker machines ref:
https://github.com/openshift/installer/pull/4456

We're now looking at other approaches to resolve the
network-config requirements driving that work, so
switching back to the pointer config for masters seems
reasonable, particularly given this issue discovered for
hive deployments.

Conflicts:
  pkg/tfvars/baremetal/baremetal.go

This reverts commit 98dc38170a66d3ede7513fc01e3ccce4ff8dd93d.