Azure: Check Azure disk Instance Type for PremiumIO Capabilities
If the Azure disktype is Premium_LRS, a check must be made to see
if the given InstanceType has the PremiumIO capabilities which when
not present causes creation of clusters to fail. Adding the check
which iterates through all the capabilities of the given instance
type disk to find the PremiumIO, check if it is set to True and
return error otherwise.