aws: update CloudFormation templates
Update the CloudFormation templates to create the infrastructure in a
way that matches how the IPI install would create the infrastructure. By
having these match by default, this gives the user a common starting
place before they make any customizations that are required during their
UPI install.

The bulk of these changes are related to ensuring that objects created
are named and tagged in the way that OpenShift expects them to be. This
will enable things such as MachineSets to work out of the box (and even
at install time for the default set of worker nodes).

These changes also include a bit of formatting update to provide
consistency across the template files and for readability.

Signed-off-by: Jared Hocutt <jhocutt@redhat.com>