Bug 1937460: Make failure to fetch cloud info non-fatal
It is possible that under some circumstance we fail to gather the
necessary information to validate the underlying cloud meets the
requirements. Recently, this was a neutron bug [1] that caused it to
return an unexpected variable type.

Our inability to perform pre-flight validation shouldn't prevent the
installer from running.

This commit makes the installer more tolerant to failures to run
validations, for OpenStack platform only. Legitimate validation
failures still block the installation.

[1] https://bugs.launchpad.net/neutron/+bug/1918565