Add `openshift-install coreos print-stream-json` command
Now that we have a standardized JSON format for the bootimages,
add a CLI command to print it.  This is necessary for UPI installs;
previously what we've done for UPI is completely ad-hoc, such
as having the docs team copy the AMI list into HTML format.  Our
UPI CI automation copies `rhcos.json` from the installer git into
a container, but customers can't do that.

In the past I've even heard of people e.g. searching in AWS
for "RHEL CoreOS" and picking AMIs based on that, which is
extremely bad because they may not even get something from
us.

This closes a major gap in allowing customers to automate UPI
installs in the same way IPI works.

One aside: because the JSON data is embedded in the installer
binary, which is in turn embedded in the release image, which
is GPG signed, one can transitively verify the integrity
via that signature.