aws: ensure default resource created for vpcs are tagged
When an AWS VPC is created, a default network ACL, route table,
and security group are created for the VPC. The installer
has not been tagging those resources. This commit adds to
code in the terraform to add all of the tags to those three
resources after the VPC is created.

Since the network acl is now being tagged, the AWS destroyer
needs to know how to handle network acls. This commit adds
code to delete network acls, skipping the default one since
it cannot be deleted. Ultimately, there are no additional
deletions performed by the destroyer as the network acl that
is newly tagged is the default for the VPC and so is skipped.

https://issues.redhat.com/browse/CORS-1657