aws: add provision checks for byo hosted zone
Add validation in the "Platform Provisioning Check" asset for the
user-provided internal hosted zone. The validation checks that
(1) the hosted zone is associated with the user-provided VPC and that
(2) the hosted zone does not contain any record sets for subdomains
of the cluster's domain.

The latter of these checks is meant to provide a modicum of protection
against the user accidentally trying to install again a cluster that
they have already installed. When it comes time to destroy the second
failed installation, the destroyer will not be able to tell that the
records sets in the hosted zone are actually being used by a
different cluster.