aws: support for cleaning up recordsets in byo hosted zone
When the user provides the private hosted zone to use for the cluster,
the destroyer still needs to delete the recordsets for the cluster
when the cluster is destroyed. There is no way to tag recordsets, so
we must rely on tagging the hosted zone as shared. When the destroyer
encounters a hosted zone tagged as shared by the cluster, the destroyer
will delete all recordsets in that hosted zone that are strict
subdomains of the cluster's domain. The cluster domain is added to the
AWS cluster metadata so that it is available to the destroyer.