images: Add Dockerfile.upi.ci.rhel8
Fork the current rhel7 Dockerfile and port it to rhel8.  This
makes it easier to use both python2 and python3 at the same time,
which we need for gcloud (python2) and azure (python3).

We need to land this, then try changing CI to actually use
it in a rehearsal job in openshift/release.