azure: changes the way we get credentials
Change in the function signatures allows a caller
to provide credentials.

This is useful when we want to avoid a dependency on a file system
or asking for manual input. This can happen in managed OpenShift services,
where some packages from the installer are being used
in the context of the webserver.