assets: use aws byo hosted zone in dns config
When the user provides an existing hosted zone to use for the
cluster's private zone, the dns.config.openshift.io resource
needs to be adjusted so that the operator can locate the
hosted zone. Since the hosted zone already exists, we can specify
the ID of the hosted zone rather than relying on tags.