installer-create: Provide user friendly error messages during failures
Refines the error messages that are provided during cluster installation
if the API or the bootstrap fails to come up within the given time
to provide more info to the user about what would have likely
caused the error. This would provide a suggestion to the user
about where to go look for the specific log file for the error.