remove ovirt_cafile from ovirt-credentials secret
When the user tried to use an ovirt.config with a ca_file the ca_file path is written to the secret but this will not be the path of the file in the created machines which will cause connection errors in the machine object that tries to find the file that doesn't exist.
Instead we should read the file content and write it as the ovirt_ca_bundle in case the ovirt_ca_bundle is not set on the ovirt.config file.