aws: support setting IAM role in the default AWS machine platform
If the IAM role is set in the default AWS machine platform but not
in the machine platform for a specific machine pool, then the IAM
role from the default AWS machine platform should be used for that
machine pool.

These changes also fix a bug where if the AWS machine platform were
omitted from a particular machine pool, the installer would panic
when trying to find IAM roles to tag.

https://bugzilla.redhat.com/show_bug.cgi?id=1946119