terraform: Consolidate bootstrap resources with module count feature
Terraform 0.13 has a module count feature that can be set to instruct
terraform to create the exact number of copies of the module in the
cloud provider. This feature can be used to control the creation and
deletion of the bootstrap resources needed for cluster.

Adding the count field in all the bootstrap modules of all cloud
providers in order to make it easy to destroy bootstrap after
nodes have been created. This field would also make it easy to
destroy the resources without having to check for the cloud
platform and have a separate destroy logic for each of the platforms.