doc/openstack: add a note about image registry in AZ
Add a note that if we deploy a cluster in an OpenStack AZ, it's
suggested to use Cinder backend if the service is available in this AZ,
since Cinder is topology aware.
Swift isn't deployed in AZs usually, so traffic would have to go
through the link between sites, which isn't optimal in real world.

Signed-off-by: Emilien Macchi <emilien@redhat.com>