aws: Periodically send shut down requests for stuck EC2 instances
During AWS destroy cluster sequence, the instances are requested
to be shut down but they do not get terminated for a long time.
The installer ignores all the EC2 instances that are in "shut-down"
status as they take a long time to complete but sometimes these
instances are stuck at the same status for a long time.

A fresh termination request has a more better chance at destroying
these stuck instances and hence changing the code to send the
termination request even for the instances in "shut-down" status.