libvirt: ALlow duplicate dnsmasq options
A corner case was discovered where if a sepcific option is specified multiple times(for example the
"address" option could be specified multiple times for different domains), only one of the options is
set. This is a bug and an oversight when adding the original code for the dnsmasq options. Fixed by using
a list instead of a map.