Terraform: Split into separate stages - golang files
Add targets "bootstrap", "cluster", and "compat" for invoking terraform. The
bootstrap and cluster targets unapck terraform into individual directories and
invokes terraform apply on the target specific terraform. The compat target
unpacks the compat directory, which keeps the orignal layout and functionality
of terraform for backward compatibility. For now, to use the split
functionality, the OS environment variable OPENSHIFT_INSTALL_SPLIT_TERRAFORM
must be set to a boolean value.

https://issues.redhat.com/browse/CORS-1693