terraform: support separate tf into stages
Add a Stage interface which defines a given terraform stage for a
given platform. For each stage, there is an option to include the
stage in the destroy when doing a bootstrap destroy. There is also
an option to provide a custom destroy function. For example, when
destroying the GCP bootstrap, there is a stage where the destroy
behavior is modifying the load balancers to remove the bootstrap
from the backend.

Use the Stage interface to run the terraform applies and destroys.
The outputs from each stage are used as input into subsequent stages.
This allows stages to get information that would either not be available
at all or only available via a data source. For example, with GCP there
is a stage where the health check IDs are needed. There is no data source
available in the gcp terraform provider for accessing these IDs. However,
using the output from the stage where the health checks are created,
the stage is able to get the health check IDs as input instead.

Include the IPs of the bootstrap instance and the control plane
instances in the output of the terraform stage where the instances
are created. This is used by the bootstrap gather as a consistent
way to grab the IPs as opposed to extracting the IPs from the
terraform state files.

When looking for the IPs of the bootstrap and control plane machines,
consider all of the terraform stages. For compat stages, there is
no change in behavior. For newer split stages, the gather looks through
the output files for the "bootstrap_ip" and "control_plane_ips"
outputs.

https://issues.redhat.com/browse/CORS-1693